requirements.txt
APIFlask==2.1.0
apispec==6.4.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
load-dotenv==0.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
webargs==8.4.0
Werkzeug==3.0.1
